requirements.txt
asgiref==3.8.1
channels==4.2.2
dj-database-url==3.0.0
Django==5.2.1
django-environ==0.12.0
django-jalali==7.4.0
django-jalali-date==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework-oauth==1.1.0
djangorestframework_simplejwt==5.5.0
jalali_core==1.0.0
jdatetime==5.2.0
mysqlclient==2.2.7
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
